+++ src/components/secciones/Contacto.js
@@ -2,7 +2,6 @@ import React, { Component } from "react";
 import { connect } from 'react-redux';
 import $ from "jquery";
 import Scrollchor from "react-scrollchor";
-import {Helmet} from "react-helmet";
 
 //CSS
 import "../../assets/css/Contacto.css";
@@ -10,6 +9,7 @@ import "../../assets/css/Contacto.css";
 //Componentes
 import Footer from "../Footer";
 import Header from "../Header";
+import SEO from "../elementos/SEO";
 
 //Images
 import flechaAbajo from "../../assets/img/angle-arrow-down.png";
@@ -104,17 +104,13 @@ class Contacto extends Component {
 		let descripcion = "Contáctate para consultar precios de baterias, volquetes y otros servicios.";
         return (
             <div className="contacto">
-                <Helmet>
-					<title>{title}</title>
-					
-					<meta name="description" content={descripcion} />
-					
-					<meta property="og:url" content="https://www.electrobatrosario.com/contacto" />
-					<meta property="og:title" content={title} />
-					<meta property="og:image" content="https://www.electrobatrosario.com/img/contacto.jpg" />
-					<meta property="og:image:alt" content="Volquetes Electrobat Rosario" />
-					<meta property="og:description" content={descripcion}/>
-				</Helmet>
+				<SEO 
+					title={title} 
+					description={descripcion} 
+					url="https://www.electrobatrosario.com/contacto"
+					ogImage="https://www.electrobatrosario.com/img/contacto.jpg"
+					ogImageAlt={`Contacto Electrobat`}
+				/>
                 <MensajeEmail
                     key="modal-2"
                     onHide={this.onHide.bind(this)}