+++ src/components/secciones/Volquetes.js
@@ -2,7 +2,6 @@ import React, { Component } from "react";
 import { connect } from "react-redux";
 import history from "../../history";
 import Scrollchor from "react-scrollchor";
-import {Helmet} from "react-helmet";
 
 //CSS
 import "../../assets/css/Volquetes.css";
@@ -10,6 +9,7 @@ import "../../assets/css/Volquetes.css";
 //Components
 import Header from "../Header";
 import Footer from "../../components/Footer";
+import SEO from "../elementos/SEO";
 
 
 //Images
@@ -40,17 +40,13 @@ class Volquetes extends Component {
 		let descripcion = title + ". Rosario y alrededores. Tamaño mediano y grande.";
         return (
             <div className="volquetes">
-                <Helmet>
-					<title>{title}</title>
-					
-					<meta name="description" content={descripcion} />
-					
-					<meta property="og:url" content="https://www.electrobatrosario.com/volquetes" />
-					<meta property="og:title" content={title} />
-					<meta property="og:image" content="https://www.electrobatrosario.com/img/volquete.jpg" />
-					<meta property="og:image:alt" content="Volquetes Electrobat Rosario" />
-					<meta property="og:description" content={descripcion}/>
-				</Helmet>
+				<SEO 
+					title={title} 
+					description={descripcion} 
+					url="https://www.electrobatrosario.com/volquetes"
+					ogImage="https://www.electrobatrosario.com/img/volquete.jpg"
+					ogImageAlt="Volquetes Electrobat Rosario"
+				/>
                 <div className="imagenVolquetes imagenFondo">
                     <Header />
                     <div className="item"></div>