+++ CHANGELOG.md
@@ -18,7 +18,7 @@ y este proyecto adhiere a [Semantic Versioning]( https://semver.org/spec/v2.0.0.
 -
 
 ### Fixed
--
+- Cambios menores index.js.
 
 ## [1.2.2] 24-01-2021
 ### Fixed